Register mapping with postgresql provider to support indexing
